+++ dzVents/runtime/tests/testDevice.lua
@@ -65,6 +65,7 @@ local function getDevice_(
 		["name"] = name,
 		["description"] = "Description 1",
 		["batteryLevel"] = batteryLevel and batteryLevel or 50,
+		["protected"] = true,
 		["signalLevel"] = signalLevel and signalLevel or 55,
 		["deviceType"] = type and type or "someSubType",
 		["deviceID"] = "123abc",
@@ -79,6 +80,7 @@ local function getDevice_(
 		["hardwareType"] = hardwareType,
 		["hardwareTypeValue"] = hardwaryTypeValue,
 		["hardwareID"] = 1,
+		['protected'] = true,
 		['_nValue'] = 123,
 		['unit'] = 1
 		},
@@ -177,7 +179,7 @@ describe('device', function()
 			['changed'] = true,
 		})
 		utils = device._getUtilsInstance()
-		utils.print = function()  end
+		utils.print = function() end
 	end)
 
 	after_each(function()
@@ -209,6 +211,7 @@ describe('device', function()
 			assert.is_same('Contact', device.switchType)
 			assert.is_same(2, device.switchTypeValue)
 			assert.is_same(true, device.timedOut)
+			assert.is_same(true, device.protected)
 			assert.is_same(50, device.batteryLevel)
 			assert.is_same(55, device.signalLevel)
 			assert.is_same('sub', device.deviceSubType)
@@ -244,7 +247,7 @@ describe('device', function()
 			device.cancelQueuedCommands()
 			assert.is_same({
 				{ ['Cancel'] = { idx = 1, type = 'device' } },
-			 }, commandArray)
+			}, commandArray)
 		end)
 
 		it('should deal with percentages', function()
@@ -541,6 +544,7 @@ describe('device', function()
 				local adapterManager = device.getAdapterManager()
 
 				assert.is_same({
+					"activate",
 					"armAway",
 					"armHome",
 					"close",
@@ -565,14 +569,15 @@ describe('device', function()
 					"playFavorites",
 					"quietOff",
 					"quietOn",
+					"reset",
 					"setColor",
 					"setColorBrightness",
-					-- "setDescription",
 					"setDiscoMode",
 					"setHex",
 					"setHotWater",
 					"setHue",
 					"setKelvin",
+					"setLevel",
 					"setMode",
 					'setNightMode',
 					'setRGB',
@@ -622,6 +627,123 @@ describe('device', function()
 			end)
 		end)
 
+		it('should have generic methods', function()
+
+			local device = getDevice(domoticz, {
+				['name'] = 'myDevice',
+				['type'] = 'Any',
+			})
+
+			assert.is_function(device.protectionOff)
+			assert.is_function(device.protectionOn)
+			assert.is_function(device.setDescription) 
+			assert.is_function(device.setIcon)
+			assert.is_function(device.setValues)
+			assert.is_function(device.rename)
+
+		end)
+
+		it('should handle generic method protectionOff ', function()
+
+			local device = getDevice(domoticz, {
+				['name'] = 'myDevice',
+				['type'] = 'Any',
+			})
+
+			local res;
+
+			domoticz.openURL = function(url)
+				res = url;
+			end
+
+			device.protectionOff()
+			assert.is_same('http://127.0.0.1:8080/json.htm?type=setused&used=true&protected=false&idx=1', res)
+		end)
+
+		it('should handle generic method protectionOn ', function()
+
+			local device = getDevice(domoticz, {
+				['name'] = 'myDevice',
+				['type'] = 'Any',
+			})
+
+			local res;
+			domoticz.openURL = function(url)
+				res = url;
+			end
+
+			device.protectionOff()
+			assert.is_same('http://127.0.0.1:8080/json.htm?type=setused&used=true&protected=false&idx=1', res)
+		end)
+
+		it('should handle generic method setDescription ', function()
+
+			local device = getDevice(domoticz, {
+				['name'] = 'myDevice',
+				['type'] = 'Any',
+			})
+
+			local res;
+
+			domoticz.openURL = function(url)
+				res = url;
+			end
+
+			device.setDescription('This is a wonderful description of the subject. Thanks !')
+			assert.is_same('http://127.0.0.1:8080/json.htm?description=This+is+a+wonderful+description+of+the+subject.+Thanks+%21&idx=1&name=myDevice&type=setused&used=true', res)
+		end)
+
+		it('should handle generic method setIcon', function()
+
+			local device = getDevice(domoticz, {
+				['name'] = 'myDevice',
+				['type'] = 'Any',
+			})
+
+			local res;
+
+			domoticz.openURL = function(url)
+				res = url;
+			end
+
+			device.setIcon(24)
+			assert.is_same('http://127.0.0.1:8080/json.htm?type=setused&used=true&name=myDevice&description=Description+1&idx=1&switchtype=2&customimage=24', res)
+		end)
+
+		it('should handle generic method rename ', function()
+
+			local device = getDevice(domoticz, {
+				['name'] = 'myDevice',
+				['type'] = 'Any',
+			})
+
+			local res;
+
+			domoticz.openURL = function(url)
+				res = url;
+			end
+
+			device.rename('who am I')
+			assert.is_same('http://127.0.0.1:8080/json.htm?type=command&param=renamedevice&idx=1&name=who+am+I', res)
+		end)
+
+		it('should handle generic method setValues', function()
+
+			local device = getDevice(domoticz, {
+				['name'] = 'myDevice',
+				['type'] = 'Any',
+			})
+
+			local res;
+
+			domoticz.openURL = function(url)
+				res = url;
+			end
+			local test = 'sValue'
+			device.setValues(12,test,13,14,test)
+			assert.is_same('http://127.0.0.1:8080/json.htm?type=command&param=udevice&idx=1&nvalue=12&svalue=sValue;13;14;sValue', res)
+		end)
+
 		it('should detect an evohome device', function()
 
 			local device = getDevice(domoticz, {
@@ -642,7 +764,7 @@ describe('device', function()
 
 			assert.is_same({ { ['SetSetPoint:1'] = '14#Permanent#2016-04-29T06:32:58Z'} }, commandArray)
 		end)
-		
+
 		it('should detect an evohome hotWater device', function()
 
 			local device = getDevice(domoticz, {
@@ -885,16 +1007,12 @@ describe('device', function()
 			assert.is_same(123.44, device.counterToday)
 			assert.is_same(6.7894, device.counter)
 
-			domoticz.openURL = function(url)
-				res = url;
-			end
-
 			device.updateCounter(555)
 			assert.is_same({ { ["UpdateDevice"] = {idx=1, nValue=0, sValue="555", _trigger=true} } }, commandArray)
-		  
-			local domoticz = require('Domoticz')
+
+			commandArray = {}
 			device.incrementCounter(10)
-			assert.is_same('http://127.0.0.1:8080/json.htm?type=command&param=udevice&idx=1&svalue=10', res)
+			assert.is_same({ { ["UpdateDevice"] = {idx=1, nValue=0, sValue="10", _trigger=true} } }, commandArray)
 		end)
 
 		it('should detect a pressure device', function()
@@ -1171,6 +1289,40 @@ describe('device', function()
 				assert.is_same({ { ["s1"] = "Set Level 50" } }, commandArray)
 			end)
 
+			it('should setLevel', function()
+				switch.setLevel(51)
+				assert.is_same({ { ["s1"] = "Set Level 51" } }, commandArray)
+			end)
+
+			it('should switch a smoke_detector', function()
+				local device = getDevice(domoticz, {
+					['type'] = 'Light/Switch',
+					['state'] = 'Off',
+					['name'] = 'mySmokeDetector',
+					['additionalRootData'] = { ['switchType'] = 'Smoke Detector'},
+				})
+				assert.is_false(device.active)
+				device.activate()
+				assert.is_same({ {["mySmokeDetector"] = "On"}}, commandArray)
+
+				local device = getDevice(domoticz, {
+					['type'] = 'Light/Switch',
+					['state'] = 'On',
+					['name'] = 'mySmokeDetector',
+					['additionalRootData'] = { ['switchType'] = 'Smoke Detector'},
+				})
+
+				assert.is_true(device.active)
+				local commandArray = {}
+
+				domoticz.openURL = function(url)
+					return table.insert(commandArray, url)
+				end
+
+				device.reset()
+				assert.is_same({ 'http://127.0.0.1:8080/json.htm?type=command&param=resetsecuritystatus&idx=1&switchcmd=Normal' }, commandArray)
+			end)
+
 			it('should switch a selector with numeric level', function()
 				local switch = getDevice(domoticz, {
 					['type'] = 'Light/Switch',
@@ -1190,7 +1342,7 @@ describe('device', function()
 					assert.is_same({ {["s1"]="Set Level 190"}}, commandArray)
 				commandArray = {} ;switch.switchSelector(-10.5)
 					assert.is_same({ {["s1"]="Set Level 0"}}, commandArray)
-				
+
 			end)
 
 			it('should switch a selector with levelname', function()
@@ -1223,7 +1375,7 @@ describe('device', function()
 						}
 				})
 				assert.is_same(10, switch.level)
-				assert.is_same({'Off', 'bb', 'cc'},  _.values(switch.levelNames))
+				assert.is_same({'Off', 'bb', 'cc'}, _.values(switch.levelNames))
 
 			end)
 
@@ -1264,6 +1416,45 @@ describe('device', function()
 				assert.is_false(scene.isSecurity)
 			end)
 
+			it('should have generic functions for scenes', function()
+				local scene = getDevice(domoticz, {
+					['baseType'] = 'scene',
+					['name'] = 'myScene',
+				})
+					assert.is_function(scene.protectionOff)
+					assert.is_function(scene.protectionOn)
+					assert.is_function(scene.setDescription) 
+					assert.is_function(scene.setIcon)
+					assert.is_function(scene.setValues)
+					assert.is_function(scene.rename)
+			end)
+
+			it('should handle generic methods for scenes', function()
+				local scene = getDevice(domoticz, {
+					['baseType'] = 'scene',
+					['name'] = 'myScene',
+					['state'] = 'On',
+					['id'] = 1,
+				})
+				local res;
+
+				domoticz.openURL = function(url)
+					res = url;
+				end
+
+				scene.protectionOff()
+				assert.is_same('http://127.0.0.1:8080/json.htm?type=updatescene&scenetype=0&protected=false&idx=1&name=myScene&description=Description+1', res)
+
+				scene.protectionOn()
+				assert.is_same('http://127.0.0.1:8080/json.htm?type=updatescene&scenetype=0&protected=true&idx=1&name=myScene&description=Description+1', res)
+
+				scene.rename('a')
+				assert.is_same('http://127.0.0.1:8080/json.htm?type=updatescene&scenetype=0&idx=1&name=a&description=Description+1', res)
+
+				scene.setDescription('groupie')
+				assert.is_same('http://127.0.0.1:8080/json.htm?type=updatescene&scenetype=0&idx=1&name=myScene&description=groupie', res)
+			end)
+
 			-- subdevices are tested in testDomoticz
 
 			it('should detect a group', function()
@@ -1275,7 +1466,7 @@ describe('device', function()
 
 				assert.is_false(group.isHTTPResponse)
 				assert.is_false(group.isVariable)
-				assert.is_false(group.isTimer)
+				assert.is_false(group.isTimer) 
 				assert.is_false(group.isScene)
 				assert.is_false(group.isDevice)
 				assert.is_true(group.isGroup)
@@ -1296,6 +1487,44 @@ describe('device', function()
 				assert.is_same({ { ['Group:myGroup'] = 'Off' } }, commandArray)
 
 			end)
+
+			 it('should have generic functions for groups', function()
+				local group = getDevice(domoticz, {
+					['baseType'] = 'group',
+					['name'] = 'myGroup',
+					['state'] = 'On'
+				})
+					assert.is_function(group.protectionOff)
+					assert.is_function(group.protectionOn)
+					assert.is_function(group.setDescription) 
+					assert.is_function(group.rename)
+			end)
+
+			it('should handle generic methods for group', function()
+				local group = getDevice(domoticz, {
+					['baseType'] = 'group',
+					['name'] = 'myGroup',
+					['state'] = 'On',
+					['id'] = 1,
+				})
+				local res;
+
+				domoticz.openURL = function(url)
+					res = url;
+				end
+
+				group.protectionOff()
+				assert.is_same('http://127.0.0.1:8080/json.htm?type=updatescene&scenetype=1&protected=false&idx=1&name=myGroup&description=Description+1', res)
+
+				group.protectionOn()
+				assert.is_same('http://127.0.0.1:8080/json.htm?type=updatescene&scenetype=1&protected=true&idx=1&name=myGroup&description=Description+1', res)
+
+				group.rename('a')
+				assert.is_same('http://127.0.0.1:8080/json.htm?type=updatescene&scenetype=1&idx=1&name=a&description=Description+1', res)
+
+				group.setDescription('groupie')
+				assert.is_same('http://127.0.0.1:8080/json.htm?type=updatescene&scenetype=1&idx=1&name=myGroup&description=groupie', res)
+			end)
 		end)
 
 		it('should detect a huelight', function()
@@ -1313,7 +1542,6 @@ describe('device', function()
 			commandArray = {}
 			device.switchOn()
 			assert.is_same({ { ["myHue"] = "On" } }, commandArray)
-
 
 			commandArray = {}
 			device.switchOff()
@@ -1347,16 +1575,16 @@ describe('device', function()
 		end)
 
 		describe('Quiet device ( quietOn and quietOff', function()
-	
+
 			local commandArray = {}
 			domoticz.openURL = function(url)
 				return table.insert(commandArray, url)
 			end
-			
+
 			local device = getDevice(domoticz, {
 				['name'] = 'quietDevice',
 				['state'] = 'On',
-					  ['type'] = 'Light/Switch',
+					['type'] = 'Light/Switch',
 				['subType'] = 'RGBWW',
 				['type'] = 'Color Switch'
 			})
@@ -1377,8 +1605,7 @@ describe('device', function()
 		describe('RGBW device #RGB', function()
 
 			local commandArray = {}
-			-- local utils = require('Utils')
-			domoticz.utils  = utils
+			domoticz.utils = utils
 
 			domoticz.openURL = function(url)
 				return table.insert(commandArray, url)
@@ -1441,29 +1668,29 @@ describe('device', function()
 
 			it('should handle setColor method with wrong values correctly' , function()
 				commandArray = {}
-				assert.is_false(device.setColor(15,31,447))  -- Should return false because of out of range parms
+				assert.is_false(device.setColor(15,31,447)) -- Should return false because of out of range parms
 				assert.is_same({},commandArray)
 			end)
 
-			it('should handle setColor method  correctly',function()
+			it('should handle setColor method correctly',function()
 				commandArray = {}
 				assert.is_nil(device.setColor(15,31,44))
 				assert.is_same({'http://127.0.0.1:8080/json.htm?type=command&param=setcolbrightnessvalue&idx=1&brightness=100&color={"m":3,"t":0,"cw":0,"ww":0,"r":15,"g":31,"b":44}'},commandArray)
 			end)
 
-			it('should handle setColorBrightness method  correctly',function()
+			it('should handle setColorBrightness method correctly',function()
 				commandArray = {}
 				assert.is_nil(device.setColorBrightness(15,31,44))
 				assert.is_same({'http://127.0.0.1:8080/json.htm?type=command&param=setcolbrightnessvalue&idx=1&brightness=100&color={"m":3,"t":0,"cw":0,"ww":0,"r":15,"g":31,"b":44}'},commandArray)
 			end)
 
-			it('should handle setDiscomode method  correctly',function()
+			it('should handle setDiscomode method correctly',function()
 				commandArray = {}
 				device.setDiscoMode(8)
 				assert.is_same({ 'http://127.0.0.1:8080/json.htm?param=discomodenum8&type=command&idx=1' }, commandArray)
 			end)
 
-			it('should handle setHue method  correctly',function()
+			it('should handle setHue method correctly',function()
 				commandArray = {}
 				assert.is_nil(device.setHue(180,11,true))
 				assert.is_same({ 'http://127.0.0.1:8080/json.htm?type=command&param=setcolbrightnessvalue&idx=1&brightness=11&hue=180&iswhite=true' }, commandArray)
@@ -1476,8 +1703,6 @@ describe('device', function()
 				assert.is_false(device.setHue(180,11,"white"))
 				assert.is_same({}, commandArray)
 			end)
-
-
 
 			it('should handle get Device with type color Switch correctly',function()
 				device = getDevice(domoticz, {
@@ -1507,10 +1732,6 @@ describe('device', function()
 				})
 			end)
 		end)
-
-
-
-
 
 		describe('Kodi', function()
 
@@ -1629,11 +1850,9 @@ describe('device', function()
 		assert.is_true(device.bState)
 		assert.is_true(device.active)
 
-
 		device = getDevice_(domoticz, 'myDevice', '', false)
 		assert.is_false(device.bState)
 		assert.is_false(device.active)
-
 
 		device = getDevice_(domoticz, 'myDevice', 'On', false)
 		assert.is_true(device.bState)
@@ -1670,6 +1889,5 @@ describe('device', function()
 			assert.is_same({{["UpdateDevice"]={idx=1, nValue=1, sValue="2", protected=true}}}, commandArray)
 		end)
 	end)
-
 
 end)