+++ dzVents/runtime/tests/testUtils.lua
@@ -52,7 +52,7 @@ describe('event helpers', function()
 
 			utils.log('something')
 
-			assert.is_same('Info:  something', printed)
+			assert.is_same('Info: something', printed)
 		end)
 
 		it('shoud not log above level', function()
@@ -78,6 +78,31 @@ describe('event helpers', function()
 
 	it('should return true if a file exists', function()
 		assert.is_true(utils.fileExists('testfile'))
+	end)
+
+	it('should right pad a string', function()
+		assert.is_same(utils.rightPad('string',7),'string ')
+		assert.is_same(utils.rightPad('string',7,'@'),'string@')
+		assert.is_same(utils.rightPad('string',2),'string')
+	end)
+    
+	it('should left pad a string', function()
+		assert.is_same(utils.leftPad('string',7),' string')
+		assert.is_same(utils.leftPad('string',7,'@'),'@string')
+		assert.is_same(utils.leftPad('string',2),'string')
+	end)
+    
+    
+	it('should center and pad a string', function()
+		assert.is_same(utils.centerPad('string',8),' string ')
+		assert.is_same(utils.centerPad('string',8,'@'),'@string@')
+		assert.is_same(utils.centerPad('string',2),'string')
+	end)
+    
+        
+	it('should pad a number with leading zeros', function()
+		assert.is_same(utils.leadingZeros(99,3),'099')
+		assert.is_same(utils.leadingZeros(999,2),'999')
 	end)
 
 	it('should return nil for os.execute (echo)', function()