+++ dzVents/documentation/history.md
@@ -1,3 +1,28 @@
+[2.4.29]
+- Add error message including affected module when module got corrupted on disk.
+- Add setLevel method for switchTypes.
+- Increased resilience against badly formatted type Time user-variables.
+- Use native domoticz command for increaseCounter method.
+- Set inverse of "set color" to Off to enable use of toggleSwitch for RGB type of devices.
+
+[2.4.28]
+- Add deviceExists(), groupExists(), sceneExists(), variableExists(), cameraExists() methods
+- increased httpResponse resilience against different use of Upper-, Lowercase in headers['content-type'] to ensure JSON conversion to Lua table
+
+[2.4.27]
+
+- Add protect attribute for devices / scenes and groups
+- Add methods protectOn and protectOff for devices / scenes and groups
+- Add functions rightPad, leftPad, centerPad, leadingZeros, numDecimals in utils
+
+[2.4.26]
+- Add Smoke detector
+
+[2.4.25]
+- Add rawDateTime
+- fix for combined device / civil[day|night]time trigger rule 
+- fix for checkFirst on stopped status
+
 [2.4.24]
 - Add method rename for devices, user-variables , scenes and groups